sia_calificaciones_ms/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4